Rename broken feed helper to useBrokenDirectCompanyCardFeedsForAdmin
Convert to a hook under TimeSensitiveSection/hooks and fix test lint/typing.
